feat(dashboard): enable SSR with loading skeleton
