+++ src/app/dashboard/loading.tsx
@@ -1,0 +1,5 @@
+import DashboardLoading from "@/components/dashboard/DashboardLoading";
+
+export default function Loading() {
+  return <DashboardLoading />;
+}
